+++ src/App.vue
@@ -1,119 +1,31 @@
 <template>
   <section>
-    <select name="" v-model="qtyCard">
-      <option value="" disabled selected>Select a option...</option>
-      <option value="1">Solo</option>
-      <option value="2">Dueto</option>
-      <option value="3">Trio</option>
-      <option value="4">Quarteto</option>
-    </select>
-    <select name="" v-model="qtyCharacter">
-      <option value="" disabled selected>Select a option...</option>
-      <option value="4">4</option>
-      <option value="5">5</option>
-      <option value="6">6</option>
-      <option value="7">7</option>
-    </select>
-
-    <button @click="startGame">Gerar</button>
-
+    <FormNewGameVue />
   </section>
 
-  <section><input type="text" v-model="guess" /> <button @click="tryToGuess">Enviar</button></section>
+  <section>
+    <FormGuess />
+  </section>
 
   <section>
-    <div v-for="(tries, indexTries) in game.tries" :key="indexTries" style="width:25%; float:left">
-      <p>{{ game.words[indexTries] }}</p>
-      <span v-for="(trye, indexTry) in tries" :key="indexTry">
-        <span class="trye" :class="trye.status">{{ trye.character }}</span>
-        <div v-if="(indexTry + 1) % qtyCharacter == 0">
-          <br>
-        </div>
-      </span>
-    </div>
-
+    <TableGuess />
   </section>
 </template>
 
 <script>
-import dictionary from "./data/dictionary.json";
+import FormNewGameVue from '@/components/FormNewGame.vue';
+import FormGuess from '@/components/FormGuess.vue';
+import TableGuess from './components/TableGuess.vue';
 export default {
   name: 'App',
   components: {
-  },
-  mounted() {
-    this.startGame();
+    FormNewGameVue,
+    FormGuess,
+    TableGuess
   },
   data() {
-    return {
-      dictionary,
-      guess: '',
-      possibilities: [],
-      qtyCard: "4",
-      qtyCharacter: "5",
-      game: {}
-    }
+    return {}
   },
-  methods: {
-    setGame() {
-      this.game = {
-        words: [
-
-        ],
-        tries: [
-
-        ],
-      }
-
-      for (let index = 0; index < this.qtyCard; index++) {
-        this.game.tries.push([])
-      }
-    },
-    getRandom() {
-      return this.possibilities[Math.floor(Math.random() * this.possibilities.length)];
-    },
-    startGame() {
-      this.setGame();
-      this.possibilities = this.dictionary.filter((word) => {
-        return word.length == this.qtyCharacter
-      });
-
-      while (this.game.words.length < this.qtyCard) {
-        let word = this.getRandom();
-        this.game.words.push(word.split(''))
-      }
-    },
-
-    tryToGuess() {
-      let word = this.guess.toUpperCase().split('');
-      const wordsToGuess = JSON.parse(
-        JSON.stringify(this.game.words)
-      );
-
-      for (let indexCard = 0; indexCard < this.qtyCard; indexCard++) {
-        let wordToGuess = wordsToGuess[indexCard]
-
-        for (let index = 0; index < word.length; index++) {
-          let cardCharacter = wordToGuess[index];
-          let guessCharacter = word[index];
-
-          let result = {
-            character: guessCharacter,
-            status: 'wrong',
-          }
-
-          if (cardCharacter == guessCharacter) {
-            result.status = 'correct'
-          }
-          else if (wordToGuess.includes(guessCharacter)) {
-            result.status = 'maybe'
-          }
-          this.game.tries[indexCard].push(result)
-        }
-      }
-      this.guess = ''
-    }
-  }
 }
 </script>
